Github logic added.
If we want to persist the environmental variables
over different steps in GitHub actions we need to
set them a bit differently then you would do on
a local computer. The need to be appended to the
$GITHUB_ENV file

The new PythonVirtualEnvironment python_requires
file creates a shell script which outputs the
following statement for each env variable
```
echo "ENVNAME=VALUE" >> $GITHUB_ENV
```

this should update the variables for the subsequent
steps. What I dislike about this fix is that it
puts pipeline logic in the conanfile, which
I don't think is the correct demarcation.

We should revisit this when we actually will work on
the epic.

Contributes to CURA-9177